+++ knappsack-core/src/main/java/com/sparc/knappsack/security/MemCachedSingleUseTokenRepositoryImpl.java
@@ -6,10 +6,11 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.util.StringUtils;
 
 import javax.annotation.PreDestroy;
 
-public class MemCachedSingleUseTokenRepositoryImpl implements SingleUseTokenRepository {
+public class MemCachedSingleUseTokenRepositoryImpl extends AbstractSingleUseTokenRepository implements SingleUseTokenRepository {
     private static final Logger log = LoggerFactory.getLogger(MemCachedSessionRegistryImpl.class);
 
     @Qualifier("memcachedClient")
@@ -21,24 +22,57 @@ public class MemCachedSingleUseTokenRepositoryImpl implements SingleUseTokenRepo
 
     @Override
     public void putToken(SingleUseToken token) {
+        if (token == null) {
+            log.error("Attempted to put null token.");
+            return;
+        } else if (!StringUtils.hasText(token.getSessionIdHash())) {
+            log.error("Attempted to put token with empty key.");
+        }
+
         client.set(token.getSessionIdHash(), tokenValiditySecond, token);
     }
 
     @Override
     public void updateToken(SingleUseToken token) {
+        if (token == null) {
+            log.error("Attempted to update with null token.");
+            return;
+        }
         client.replace(token.getSessionIdHash(), tokenValiditySecond, token);
     }
 
     @Override
     public void removeToken(String key) {
+        if (!StringUtils.hasText(key)) {
+            log.error("Attempted to remove a token with an empty key.");
+            return;
+        }
         client.delete(key);
     }
 
     @Override
     public SingleUseToken getToken(String key) {
+        if (!StringUtils.hasText(key)) {
+            log.error("Attempted to get a token for an empty key.");
+            return null;
+        }
         SingleUseToken token = (SingleUseToken) client.get(key);
         removeToken(key);
         return token;
+    }
+
+    @Override
+    public boolean validateAndExpireTokenForKey(String key, boolean shouldExpireToken) {
+        if (!StringUtils.hasText(key)) {
+            log.error("Attempted to validate and expire a token for an empty key.");
+            return false;
+        }
+
+        if (shouldExpireToken) {
+            return isValidToken(getToken(key), tokenValiditySecond);
+        } else {
+            return isValidToken((SingleUseToken) client.get(key), tokenValiditySecond);
+        }
     }
 
     @PreDestroy
